refactor: 패키지명 변경 entity -> domain
entity 라는 패키지명보다는 domain (해결하고자하는 문제 혹은 관심사)이 보다 적절하므로 수정하였습니다.
